reference/oap_agent/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "oap-agent"
version = "0.1.0"
description = "OAP Agent — chat + autonomous task execution via OAP discovery"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "httpx>=0.28",
    "pyyaml>=6.0",
    "apscheduler>=3.10,<4",
    "croniter>=1.3",
]

[project.scripts]
oap-agent-api = "oap_agent.api:main"
